completed restructurization of nosql.py and created subconnectors/nosql/mongodb.py subconnector for request
